Fix bug with search bar on mobile
The sidebar script was loading for some reason, which meant that it would try to find the aside and sidebar with getElementById, but that would fail since they didn't exist on the homepage. This resulted in nothing loading on mobile, since the sidebar script would run before the search script.