Common package (#45)
* Create common package

* Use Head from common in blog

* Use Head from common in gallery

* Use Head from common in homepage

* Create common styles

* Add common package to root README

* Update stylesheet import in common README

* Update head options in common README

* Create common footer

* Create common 404 page

* Create common sidebar

* Add changesets

* Add input to Inter font

* Make nav img always center aligned

* Fix bug with search bar on mobile

The sidebar script was loading for some reason, which meant that it would try to find the aside and sidebar with getElementById, but that would fail since they didn't exist on the homepage. This resulted in nothing loading on mobile, since the sidebar script would run before the search script.

* Create common package

* Use Head from common in blog

* Use Head from common in gallery

* Use Head from common in homepage

* Create common styles

* Add common package to root README

* Update stylesheet import in common README

* Update head options in common README

* Create common footer

* Create common 404 page

* Create common sidebar

* Add changesets

* Add input to Inter font

* Make nav img always center aligned

* Fix bug with search bar on mobile

The sidebar script was loading for some reason, which meant that it would try to find the aside and sidebar with getElementById, but that would fail since they didn't exist on the homepage. This resulted in nothing loading on mobile, since the sidebar script would run before the search script.